requirements.txt
environs==8.0.0
python-telegram-bot==12.5.*
requests==2.28.*
